refactor: consolidate progressView logging and inline event handlers
Major refactoring to reduce abstraction overhead in progressView:

Logger consolidation:
- Create shared progressViewLogger instance in progressViewLogger.ts
- Replace 9+ independent AgentLogger instances across managers
- Reduces logger instantiation overhead and improves consistency

Event handler simplification:
- Inline all event handlers directly in ProgressEventHandler
- Create handlers.ts with consolidated handler functions
- Delete 8 factory module files (LogEvents, OutputEvents, etc.)
- Remove unnecessary factory pattern indirection

Files deleted:
- ApprovalEvents.ts, LogEvents.ts, OutputEvents.ts, RetryEvents.ts
- StreamStatusEvents.ts, TaskGroupEvents.ts, TodoEvents.ts, UsageEvents.ts

Files added:
- progressViewLogger.ts - shared logger instance
- handlers.ts - consolidated event handler functions

Code reduction: ~800 lines of factory boilerplate removed
