+++ src/progressView/events/index.ts
@@ -1,28 +1,8 @@
 // Barrel export for progress view events
 export { withEventErrorHandling } from './errorHandling';
-export { LogEventsModule, createLogEvents } from './LogEvents';
-export { OutputEventsModule, createOutputEvents } from './OutputEvents';
-export { ProgressEventHandler } from './ProgressEventHandler';
 export {
-  RetryEventsModule,
-  RetryEventsShared,
-  createRetryEvents,
-} from './RetryEvents';
-export {
-  ApprovalEventsModule,
-  ApprovalEventsShared,
-  createApprovalEvents,
-} from './ApprovalEvents';
-export {
-  StreamStatusEventShared,
-  StreamStatusEventModule,
-  createStreamStatusEvents,
-} from './StreamStatusEvents';
-export {
-  TaskGroupEventsModule,
-  createTaskGroupEvents,
-} from './TaskGroupEvents';
-export { TodoEventsModule, createTodoEvents } from './TodoEvents';
+  ProgressEventHandler,
+  type ProgressEventHandlerCallbacks,
+} from './ProgressEventHandler';
 export { ProgressEventBusLike } from './types';
 export type { StreamStatus } from '@eventBus/ProgressEventBus';
-export { UsageEventsModule, createUsageEvents } from './UsageEvents';
