+++ src/progressView/managers/RunInstructionManager.ts
@@ -3,12 +3,12 @@ import type { StorageKey, StreamTabId } from '@agent/types/IdentifierTypes';
 
 // Internal imports
 import { WorkspaceStateKey } from '@common/state/stateManager';
-import { AgentLogger } from '@logger/AgentLogger';
 
 // Local imports - types
 import type { InstructionUpdate } from '@progressView/types';
 
 // Internal imports
+import { progressViewLogger } from '@progressView/progressViewLogger';
 import {
   PersistentMapManager,
   type StateStorage,
@@ -24,11 +24,8 @@ export class RunInstructionManager extends PersistentMapManager<
   StreamTabId,
   InstructionMap
 > {
-  private readonly logger: AgentLogger;
-
   constructor(storage?: StateStorage) {
     super(WorkspaceStateKey.RUN_INSTRUCTIONS, storage);
-    this.logger = new AgentLogger('RunInstructionManager');
   }
 
   getInstructions(stream: StreamTabId): InstructionMap {
@@ -98,7 +95,7 @@ export class RunInstructionManager extends PersistentMapManager<
       const entries = Object.entries(data as Record<string, InstructionUpdate>);
       return new Map(entries as [string, InstructionUpdate][]);
     } catch (error) {
-      this.logger.warn(
+      progressViewLogger.warn(
         `Failed to deserialize run instructions for ${stream}: ${String(error)}`,
       );
       return new Map();
