+++ src/progressView/managers/WebviewUpdater.ts
@@ -9,12 +9,12 @@ import type { AgentTypeFilter } from '@agent/types/AgentStreamTypes';
 import type { TokenUsageStats } from '@agent/types/UsageTypes';
 
 import { type StreamStatus } from '@common/constants/streamStatus';
-import { AgentLogger } from '@logger/AgentLogger';
 import { LogMessageData } from '@logger/LogTypes';
 import type { TaskState } from '@logger/TaskState';
 import type { InstructionUpdate, StreamTabInfo } from '@progressView/types';
 // Internal imports
 import { buildStreamInfos } from '@progressView/streamInfoUtils';
+import { progressViewLogger } from '@progressView/progressViewLogger';
 
 // Internal imports
 import { ProgressViewState } from '@progressView/state/ProgressViewState';
@@ -34,11 +34,7 @@ import type { TodoItem, UpdateTaskGroupPayload } from '@eventBus/schemas';
  * without coupling business logic to DOM operations.
  */
 export class WebviewUpdater {
-  private readonly logger: AgentLogger;
-
-  constructor(private getWebview: () => vscode.Webview | undefined) {
-    this.logger = new AgentLogger('WebviewUpdater');
-  }
+  constructor(private getWebview: () => vscode.Webview | undefined) {}
 
   /** Helper to send messages to webview, eliminating repetitive null checks */
   private sendMessage(message: any): void {
@@ -371,7 +367,7 @@ export class WebviewUpdater {
 
     this.updateStreams(streams, activeStream, state.agentTypeFilter);
 
-    this.logger.debug(
+    progressViewLogger.debug(
       `Updated webview streams (${streams.length}) active: ${activeStream}`,
     );
