+++ src/test/progressView/events/LogEvents.test.ts
@@ -6,8 +6,14 @@ import type { LogMessageData } from '@logger/LogTypes';
 import { MESSAGE_TYPES } from '@logger/messageTypes';
 
 // Local imports - progress view
-import { createLogEvents } from '@progressView/events/LogEvents';
-import type { ProgressEventBusLike } from '@progressView/events/types';
+import {
+  handleAddLogMessage,
+  handleUpdateLogMessage,
+} from '@progressView/events/handlers';
+import {
+  createStatefulEventDisposable,
+  type ProgressEventBusLike,
+} from '@progressView/events/types';
 import { StreamTabsManager } from '@progressView/managers/StreamTabsManager';
 import type { WebviewUpdater } from '@progressView/managers/WebviewUpdater';
 import type { StateStorage } from '@progressView/persistence/PersistentMapManager';
@@ -78,8 +84,22 @@ describe('LogEvents', () => {
     } as unknown as WebviewUpdater;
 
     const bus = new TestBus();
-    const { register } = createLogEvents();
-    const disposables = register(bus, state, updater);
+    const disposables = [
+      createStatefulEventDisposable(
+        bus,
+        'addLogMessage',
+        state,
+        updater,
+        handleAddLogMessage,
+      ),
+      createStatefulEventDisposable(
+        bus,
+        'updateLogMessage',
+        state,
+        updater,
+        handleUpdateLogMessage,
+      ),
+    ];
 
     const thinkingMessage: LogMessageData = {
       id: 'log-1',
